Not exporting exception classes and other class hierarchies under Windows. Fixed MSVC compiler warnings.

+++ src/library/framework_full.cpp
@@ -204,7 +204,7 @@ void Full(
    ParameterError parameterError;
    RunTimeError runTimeError;
    Error error;
-   #pragma omp parallel num_threads( nThreads )
+   #pragma omp parallel num_threads( static_cast< int >( nThreads ))
    try {
       dip::uint thread = static_cast< dip::uint >( omp_get_thread_num() );
